fix deprecated warning of inspect
inspect.getargspec is deprecated since Python 3.0
